Refactored smhm_models test suite
